refactor(contract): PaymentType을 여러 도메인서 공유하도록 common으로 이동
