+++ contract-service/src/main/java/com/example/contractservice/common/PaymentType.java
@@ -1,4 +1,4 @@
-package com.example.contractservice.contract.common;
+package com.example.contractservice.common;
 
 public enum PaymentType {
     ONE_TIME,
