+++ contract-service/src/main/java/com/example/contractservice/contract/domain/vo/ContractInfo.java
@@ -1,7 +1,7 @@
 package com.example.contractservice.contract.domain.vo;
 
 import com.example.contractservice.contract.common.ContractStatus;
-import com.example.contractservice.contract.common.PaymentType;
+import com.example.contractservice.common.PaymentType;
 import java.time.Instant;
 
 public record ContractInfo(
